Stream Tart stdout/stderr output to Packer UI
Creating a Tart VM may involve long running operations such as
downloading an IPSW, or installing the OS. Tart gives us progress
during these operations, so let's forward it to the Packer UI,
the same way the shell-local provisioner does.

If a packer.Ui is not provided, we keep the previous code path
that returns the output to the caller, for example used by the
TartMachineIP helper.