Use packer plugins install --path to install development version
With modern Packer, and Packer templates with required_plugins
sections, the plugins need to follow a specific folder structure
and also come with a sha256 hash of the plugin.

By using `packer plugins install --path` we can let Packer deal
with the specifics of installing a plugin from disk instead of
from a repository.

As a drive-by, update the version number in version.go to match
the latest tagged version. This version number should always
reflect the current version.