Use same steps for Linux installation as for macOS installation
The code for Linux installation was maintained separately in
stepCreateVM, even though it shared most of the logic with
the macOS installation flow of stepRun.

We now use the same flow for Linux and macOS installation:

 1. The VM is created in stepCreateVM
 2. We run the VM in stepRun, which based on the
    `boot_command` does automated installation of
    the OS
 3. Once the automated installation is done, and SSH
    access is enabled, SSH provisioning steps can
    proceed

To achieve this flow for Linux, the autoinstall cloud
config in the Ubuntu example had to be changed to from
"poweroff" to "reboot", so that the VM was rebooted into
a level where SSH is available. This is the default value
for the autoinstall shutdown key, and the only reason we
needed to override it previously was because we did the
installation (via boot_command) as part of stepCreateVM.

The SSH timeout in the example also had to be increased,
as we now wait for the installation to be completed,
just like we for macOS.