Generate plugin version at build time based on git describe
If the git describe is a plain tag, the resulting plugin
version matches the tag.

If there are commits on top of the most recent tag, or the
working directory is dirty, the resulting plugin version is
the latest plugin version with the patch version bumped,
and the "dev" suffix added.

The same logic has been applied to the goreleaser file,
so that the releasing logic can be tested with --snapshot