Add <wait "string"> boot command that waits for text to appear
The syntax is supported without having to modify the packer-sdk
bootcommand grammar by a simple regex replacement that uses two
symbols from the Unicode Private Use Area as markers, combined
with a custom VNC driver that handles the markers in SendKey().

The actual recognition of the text is handled by Vision.framework.

To aid debugging the frames we analyze are written out as PNG
files if packer is run with -debug.